Pull request #58: Add delete message
Merge in TO2024/mi-pt-1500-jelenie from add-delete-message to Milestone_3

Squashed commit of the following:

commit 9f37815745c4d61526be8421476a0ef6f5f471c7
Author: Łukasz <lukasz.9@o2.pl>
Date:   Sat Jan 18 17:45:19 2025 +0100

    Fix bugs

commit c219bce65d8ae43323b0da5ba9ecc0125778adab
Author: Lukasz <lukasz.9@o2.pl>
Date:   Fri Jan 17 13:11:05 2025 +0100

    Attempt to fix bugs

commit bf2cf8bb9ca06b2ce14a04e4a03347056fdf787b
Merge: 2a0fbe1 b2045ac
Author: Lukasz <lukasz.9@o2.pl>
Date:   Fri Jan 17 04:25:16 2025 +0100

    Merge branch 'Milestone_3' into add-delete-message
    
    # Conflicts:
    #	backend/build/libs/oot-0.0.1-SNAPSHOT-plain.jar
    #	backend/build/libs/oot-0.0.1-SNAPSHOT.jar
    #	backend/src/main/java/agh/project/oot/messages/Message.java
    #	backend/src/main/java/agh/project/oot/service/MessageService.java

commit 2a0fbe1b1f6e74ba728c9b704be5185536666d98
Author: Lukasz <lukasz.9@o2.pl>
Date:   Fri Jan 17 04:11:21 2025 +0100

    Add delete message

+++ backend/src/main/java/agh/project/oot/messages/PlaceholderNumberMessage.java
@@ -8,9 +8,9 @@ import lombok.Setter;
 @Getter
 @Setter
 public final class PlaceholderNumberMessage extends Message {
-    private Integer thumbnailsNumber;
+    private Long thumbnailsNumber;
 
-    public PlaceholderNumberMessage(Integer thumbnailsNumber) {
+    public PlaceholderNumberMessage(Long thumbnailsNumber) {
         super(ConnectionStatus.CONNECTED, ResponseStatus.OK, MessageType.PLACEHOLDERS_NUMBER_RESPONSE);
         this.thumbnailsNumber = thumbnailsNumber;
     }
